requirements.txt
Django==5.2.6
gunicorn==21.2.0
psycopg2-binary==2.9.7
dj-database-url==2.1.0
whitenoise==6.6.0
djangorestframework==3.15.1
reportlab==4.0.7
numpy==1.24.3
scikit-learn==1.3.0
pandas==2.0.3
